* 'master' of https://github.com/donovanOng/comp3013:
  Profile page
  Leave Circle by Member and Edit Circle Name by Owner
  View friend's profile (Privacy setting not done)
  Margin bottom for Circle cards
  List of Profiles, Collections, Photos and Comment on Admin index
  1) Allows Profile to be created or updated (Update SQL not working yet) 2) Annotation table and data
  More details for tables in Admin index
  Admin edit Message content and delete message
  Admin edit circle name and delete circle
  Admin column in User table
  Indicate Owner of Circle and only Owner can remove member
  Order Circle’s messages by UPDATED_AT
  Message when user not in any circle
  Hide list of friends when user has no friend
  Padding for bottom of page
  Edit and Delete User for Admin
  Margin for friend’s card
  Replace user index with admin index
  Messages for Friendship Request actions
  Redirect to Homepage for empty user search

+++ application/view/circles/index.php
@@ -3,7 +3,7 @@
 <div class="card">
   <div class="card-block">
 
-    <div class="row align-items-center mb-3">
+    <div class="row align-items-center">
       <div class="col-10"><h4>Circles</h4></div>
       <div class="col-2 text-right">
         <? if ($user->userID == $this->current_userID) { ?>
@@ -15,39 +15,44 @@
     </div>
     <div class="mt-3 mb-3" style="border-top: 1px solid #DDD;"></div>
     <? require APP . 'view/circles/new.php'; ?>
-    <div class="row">
-      <? if (count($circlesAdmin) > 0) { ?>
-      <?php foreach ($circlesAdmin as $circle) { ?>
-        <div class="col-3">
-          <div class="card mt-1 mb-1 bg-faded">
-            <div class="card-block">
-              <h5 class="card-title"><?= $circle->name ?></h5>
-              <p class="card-text text-muted">
-                # members
-              </p>
-              <a href="<?= URL; ?>circle/<?= $circle->circleID ?>" class="card-link">View Circle</a>
-              <a href="<?= URL; ?>circle/delete?circleID=<?= $circle->circleID?>" class="card-link">Delete</a>
+    <? if ( (count($circlesAdmin) + count($circlesMember))  == 0) { ?>
+      <p class="pl-1 pr-1 m-0">Not in any Circle.</p>
+    <? } else { ?>
+      <div class="row align-items-top">
+        <? if (count($circlesAdmin) > 0) { ?>
+        <?php foreach ($circlesAdmin as $circle) { ?>
+          <div class="col-3">
+            <div class="card mb-3 bg-faded">
+              <div class="card-block">
+                <h5 class="card-title"><?= $circle->name ?></h5>
+                <p class="card-text text-muted">
+                  # members
+                </p>
+                <a href="<?= URL; ?>circle/<?= $circle->circleID ?>" class="card-link">View Circle</a>
+                <? if ($this->current_userID == $circle->userID) { ?>
+                  <a href="<?= URL; ?>circle/delete?circleID=<?= $circle->circleID?>" class="card-link">Delete</a>
+                <? } ?>
+              </div>
             </div>
           </div>
-        </div>
-      <?php } ?>
-      <? } ?>
-      <? if (count($circlesMember) > 0) { ?>
-      <?php foreach ($circlesMember as $circle) { ?>
-        <div class="col-3">
-          <div class="card mb-3">
-            <div class="card-block">
-              <h5 class="card-title"><?= $circle->name ?></h5>
-              <p class="card-text text-muted">
-                # members
-              </p>
-              <a href="<?= URL; ?>circle/<?= $circle->circleID ?>" class="card-link">View Circle</a>
-              <a href="<?= URL; ?>circle/delete?circleID=<?= $circle->circleID ?>" class="card-link">Delete</a>
+        <?php } ?>
+        <? } ?>
+        <? if (count($circlesMember) > 0) { ?>
+        <?php foreach ($circlesMember as $circle) { ?>
+          <div class="col-3">
+            <div class="card mb-3">
+              <div class="card-block">
+                <h5 class="card-title"><?= $circle->name ?></h5>
+                <p class="card-text text-muted">
+                  # members
+                </p>
+                <a href="<?= URL; ?>circle/<?= $circle->circleID ?>" class="card-link">View Circle</a>
+              </div>
             </div>
           </div>
-        </div>
-      <?php } ?>
-      <? } ?>
-    </div>
+        <?php } ?>
+        <? } ?>
+      </div>
+    <? } ?>
   </div>
 </div>
